refactor(framework): extract synthetic pipeline steps from route.ts
Moves buildParseStep, buildCacheCheckStep, buildCacheStoreStep and
their synthetic adapter carriers verbatim into pipeline/synthetic-steps.ts.
These have zero coupling to DefaultRoute. route.ts re-exports the cache
builders for existing @internal imports (builder.ts). route.ts drops
from 1,870 to 1,522 lines.
